Fix MSVC compiler warnings and platform-specific configuration
- Fix MSVC /W3 overriding warnings by using platform-specific flags
- Add explicit /W4 flag for Windows builds
- Fix size_t to int conversion warning in examples/basic_usage.cpp
- Add platform-specific compiler flag configuration to avoid conflicts
- Improve xmake.lua with better Windows/Linux compatibility

+++ examples/basic_usage.cpp
@@ -318,7 +318,7 @@ void demonstrate_eda_design_rules() {
         for (size_t i = 0; i < eda_components.size(); ++i) {
             auto sharp_angles = eda_components[i].get_sharp_angles(process.sharp_angle_limit);
             if (!sharp_angles.empty()) {
-                violations += sharp_angles.size();
+                violations += static_cast<int>(sharp_angles.size());
                 std::cout << "  " << component_names[i] << ": " << sharp_angles.size() 
                          << " sharp angle violations" << std::endl;
             }
